Remove dependency on crypto utils from the JS AWS SDK.
This enables customers to use this SDK without needing to depend on the AWS SDK.
There is not a significant impact to bundle size because browser crypto tools are leveraged and there is no more piping of the dependencies.
The downside of this change is that it restricts the SDK to only be browser compatible.